fix: Update codegen.yaml to generate clients in three batches instead of four
The current codegen.yaml generates clients in four batches, which makes the last batch always succeeds due to the number of services is less than 300 at this moment. 
Changing the batch size to three so that the build status is correctly reflected.
We need to manually increase the batch size once the number of clients is greater than 300.