fix: Update codegen.yaml to generate clients in 2 batches and 200 clients per batch (#17551)
The current codegen.yaml generates clients in 4 batches and 100 clients per batch, which makes the last batch always succeeds due to the number of services is less than 300 at this moment. Changing the batch size to 2 and number of clients per batch to 200 so that the build status is correctly reflected.